python/requirements.txt
mastodon.py==1.5.2